+++ packages/sdk-typescript/src/resources/files.ts
@@ -48,7 +48,7 @@ export class FilesResource {
         });
         return response as OperationResult<Blob>;
       },
-      { method: 'GET' }
+      { method: 'GET', responseType: 'binary' }
     );
   }